Support filtering part of the filename.
Coverity analysis may not be performed on the same directory as Sonar analysis,
so in some case we need to remove the beginning of the filename to make it
relative to Sonar's project root.

This can be done by specifying the prefix to remove from filenames with the
'sonar.coverity.prefix' key.