Fix SonarQube analysis on Windows systems.
On Windows systems the SonarQube use backslashes in the path of resources,
but Coverity always store the filepath in unix style (with slash). If the
SonarQube Runner runs on Windows, then we have to change the path to unix
path, so the SonarQube can create the SonarQube issues from the Coverity
defects.