Remove CoverityWidget
As the CoverityWidget is no longer used in 7.9 and 8.9 LTS and the data it
displayed can still be found in the Measures tab.

The CoverityWidget code will cause a ClassNotFoundException in SonarQube 8.9
LTS upon startup (while it still works in SonarQube 7.9 LTS).  There was a
major overhaul that occurred in SonarQube 6.2 that removed Widgets and
customized dashboards (see https://jira.sonarsource.com/browse/SONAR-8354).